Replace Kusama and Westend networks with Zenchain network
The Kusama and Westend networks have been removed and replaced with the Zenchain network. This is reflected across multiple files including network configuration and ledger definitions. Furthermore, a script for updating Zenchain's chainspec from the official repository was added, also a new Zenchain chainspec file was included in the current repo.